requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distlib==0.3.9
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
fsspec==2025.3.2
gunicorn==23.0.0
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
lime==0.2.0.1
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.2.1
platformdirs==4.3.7
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.15
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.30.0
Werkzeug==3.1.3
